Add configurable threshold for immediate mod activation
This feature enables the user to configure two properties with which
faster sameside or crossover combinations can be achieved.

BILATERAL_COMBINATIONS_IMMEDIATE_SAMESIDED_CHORD_SIZE_THRESHOLD
BILATERAL_COMBINATIONS_IMMEDIATE_CROSSOVER_CHORD_SIZE_THRESHOLD

The above mentioned properties configure the number of mods that
have to be simultaneously held to immediately be activated on the
next keypress.